Testing the feature SendVerificationEmail
This feature allows the checking of the email user. It sends an email to the user that just registered himself and asks him to click on a link. This link when clicked on triggers the application to perform a set of instructions that will verify that it is the owner of the email address and grants him the access to the site.